Modifying install script and organizing nvim config

+++ .scripts/symlink-install.sh
@@ -2,6 +2,8 @@
 
 SCRIPT_DIR="$( cd "$( dirname "${(%):-%x}" )" && pwd )"
 
-if ! ln -fs "$SCRIPT_DIR" ~/.config; then
-        return 1
+if sudo pacman -S --noconfirm alacritty; then
+    if ! ln -fs "$SCRIPT_DIR" ~/.config; then
+            return 1
+    fi
 fi
